Chech flag before removing ip

+++ app/services/data_centers/merge_duplicates.rb
@@ -158,16 +158,17 @@ module DataCenters
       unless val.data_center.id == validator_ip.data_center.id        
         message = <<-EOS
           Validator #{val.name} (##{val.id}) current data center 
-          is different than assigned to validator ip #{validator_ip.address} (##{validator_ip.id}).
+          is different than assigned to validator ip #{validator_ip.address} (##{validator_ip.id})
+          validator IP will be destroyed.
         EOS
 
         log_message(message)
 
-        if validator_ip.gossip_node.nil? && validator_ip.is_active == false
+        if validator_ip.gossip_node.nil? && validator_ip.is_active == false && @run_update
           address = validator_ip.address
           id = validator_ip.id
 
-          validator_ip.destroy
+          validator_ip.destroy 
 
           message = <<-EOS
             Validator IP #{address} (##{id}) destroyed.
